Seperated material.py a bit
