+++ deepdrr/material/__init__.py
@@ -1,4 +1,5 @@
-from .material import Material, CoefficientEntry
+from .material import Material
+from .coefficient import CoefficientEntry
 from .log_interp import log_interp
 
 __all__ = [
