+++ deepdrr/material/mappings.py
@@ -1,0 +1,102 @@
+element_map = {
+    "H":  "1_H_Hydrogen",
+    "He": "2_He_Helium",
+    "Li": "3_Li_Lithium",
+    "Be": "4_Be_Beryllium",
+    "B":  "5_B_Boron",
+    "C":  "6_C_Carbon, Graphite",
+    "N":  "7_N_Nitrogen",
+    "O":  "8_O_Oxygen",
+    "F":  "9_F_Fluorine",
+    "Ne": "10_Ne_Neon",
+    "Na": "11_Na_Sodium",
+    "Mg": "12_Mg_Magnesium",
+    "Al": "13_Al_Aluminum",
+    "Si": "14_Si_Silicon",
+    "P":  "15_P_Phosphorus",
+    "S":  "16_S_Sulfur",
+    "Cl": "17_Cl_Chlorine",
+    "Ar": "18_Ar_Argon",
+    "K":  "19_K_Potassium",
+    "Ca": "20_Ca_Calcium",
+    "Sc": "21_Sc_Scandium",
+    "Ti": "22_Ti_Titanium",
+    "V":  "23_V_Vanadium",
+    "Cr": "24_Cr_Chromium",
+    "Mn": "25_Mn_Manganese",
+    "Fe": "26_Fe_Iron",
+    "Co": "27_Co_Cobalt",
+    "Ni": "28_Ni_Nickel",
+    "Cu": "29_Cu_Copper",
+    "Zn": "30_Zn_Zinc",
+    "Ga": "31_Ga_Gallium",
+    "Ge": "32_Ge_Germanium",
+    "As": "33_As_Arsenic",
+    "Se": "34_Se_Selenium",
+    "Br": "35_Br_Bromine",
+    "Kr": "36_Kr_Krypton",
+    "Rb": "37_Rb_Rubidium",
+    "Sr": "38_Sr_Strontium",
+    "Y":  "39_Y_Yttrium",
+    "Zr": "40_Zr_Zirconium",
+    "Nb": "41_Nb_Niobium",
+    "Mo": "42_Mo_Molybdenum",
+    "Tc": "43_Tc_Technetium",
+    "Ru": "44_Ru_Ruthenium",
+    "Rh": "45_Rh_Rhodium",
+    "Pd": "46_Pd_Palladium",
+    "Ag": "47_Ag_Silver",
+    "Cd": "48_Cd_Cadmium",
+    "In": "49_In_Indium",
+    "Sn": "50_Sn_Tin",
+    "Sb": "51_Sb_Antimony",
+    "Te": "52_Te_Tellurium",
+    "I":  "53_I_Iodine",
+    "Xe": "54_Xe_Xenon",
+    "Cs": "55_Cs_Cesium",
+    "Ba": "56_Ba_Barium",
+    "La": "57_La_Lanthanum",
+    "Ce": "58_Ce_Cerium",
+    "Pr": "59_Pr_Praseodymium",
+    "Nd": "60_Nd_Neodymium",
+    "Pm": "61_Pm_Promethium",
+    "Sm": "62_Sm_Samarium",
+    "Eu": "63_Eu_Europium",
+    "Gd": "64_Gd_Gadolinium",
+    "Tb": "65_Tb_Terbium",
+    "Dy": "66_Dy_Dysprosium",
+    "Ho": "67_Ho_Holmium",
+    "Er": "68_Er_Erbium",
+    "Tm": "69_Tm_Thulium",
+    "Yb": "70_Yb_Ytterbium",
+    "Lu": "71_Lu_Lutetium",
+    "Hf": "72_Hf_Hafnium",
+    "Ta": "73_Ta_Tantalum",
+    "W":  "74_W_Tungsten",
+    "Re": "75_Re_Rhenium",
+    "Os": "76_Os_Osmium",
+    "Ir": "77_Ir_Iridium",
+    "Pt": "78_Pt_Platinum",
+    "Au": "79_Au_Gold",
+    "Hg": "80_Hg_Mercury",
+    "Tl": "81_Tl_Thallium",
+    "Pb": "82_Pb_Lead",
+    "Bi": "83_Bi_Bismuth",
+    "Po": "84_Po_Polonium",
+    "At": "85_At_Astatine",
+    "Rn": "86_Rn_Radon",
+    "Fr": "87_Fr_Francium",
+    "Ra": "88_Ra_Radium",
+    "Ac": "89_Ac_Actinium",
+    "Th": "90_Th_Thorium",
+    "Pa": "91_Pa_Protactinium",
+    "U":  "92_U_Uranium",
+
+    # Add named materials too
+    "soft tissue": "tissue",
+    "tissue_soft": "tissue",
+    "iron":        "26_Fe_Iron",
+    "lead":        "82_Pb_Lead",
+    "copper":      "29_Cu_Copper",
+    "titanium":    "22_Ti_Titanium",
+}
